fix: 优化grid代码结构
